Update to 0.2.3
Summary:
- Add tests for plugin decorators.
- Allow library users to properly cleanup the event loop.
- Expose the asyncio event loop as `nvim.loop`.
- MsgpackStream: fix typos.

Also a `close()` method is added on Neovim session objects. When used as
a library for externally connecting to a nvim instance (i.e. not rplugins),
it is recommended to call the `close()` method on the session object when it is
not needed anymore.

Signed-off-by: Federico Damián Schonborn <federicodamiansch@gmail.com>

Test Plan: `nvim +CheckHealth` reports the Python 2 and 3 clients are updated and working correctly.

Reviewers: #triage_team, joebonrichie

Reviewed By: joebonrichie

Differential Revision: https://dev.solus-project.com/D2421